Rename Common to Shared directory
